Copy-paste.
I'm not sure if this is a good decision, but I took a piece of the source code
placed here: http://www.linuxcenter.ru/lib/books/lkmpg.phtml#CHARDEVC. I
will use it as the backbone for this little project.
Spacing fixed, by the way (accordingly to the preferred codestyle).